@@ -21,19 +21,19 @@
   },
   "homepage": "https://github.com/n-e/totoz.eu#readme",
   "dependencies": {
-    "body-parser": "1.19.0",
+    "body-parser": "^1.19.2",
     "connect-flash": "0.1.1",
     "connect-pg-simple": "7.0.0",
     "drupal-hash": "1.0.4",
     "escape-html": "1.0.3",
-    "express": "4.16.4",
-    "express-session": "1.17.0",
+    "express": "^4.17.3",
+    "express-session": "^1.17.2",
     "fp-ts": "^2.11.9",
     "image-size": "1.0.1",
     "io-ts": "2.2.16",
     "jsdom": "19.0.0",
-    "morgan": "1.9.1",
-    "multer": "1.4.2",
+    "morgan": "^1.10.0",
+    "multer": "^1.4.4",
     "node-fetch": "2.6.7",
     "passport": "0.5.2",
     "passport-local": "1.0.0",
@@ -42,25 +42,22 @@
     "xml-js": "1.6.11"
   },
   "devDependencies": {
-    "@types/connect-flash": "0.0.34",
-    "@types/connect-pg-simple": "4.2.2",
-    "@types/escape-html": "0.0.20",
-    "@types/express": "4.16.1",
-    "@types/express-session": "1.15.12",
-    "@types/image-size": "0.0.29",
-    "@types/jsdom": "11.12.0",
-    "@types/morgan": "1.7.35",
-    "@types/multer": "1.3.7",
-    "@types/node-fetch": "2.1.6",
-    "@types/passport": "0.4.7",
-    "@types/passport-local": "1.0.33",
-    "@types/pg": "7.14.10",
-    "concurrently": "3.6.1",
-    "nodemon": "1.18.10",
+    "@types/connect-flash": "^0.0.37",
+    "@types/connect-pg-simple": "^7.0.0",
+    "@types/escape-html": "^1.0.1",
+    "@types/express": "^4.17.13",
+    "@types/express-session": "^1.17.4",
+    "@types/image-size": "^0.8.0",
+    "@types/jsdom": "^16.2.14",
+    "@types/morgan": "^1.9.3",
+    "@types/multer": "^1.4.7",
+    "@types/node-fetch": "^2.6.1",
+    "@types/passport": "^1.0.7",
+    "@types/passport-local": "^1.0.34",
+    "@types/pg": "^8.6.5",
+    "concurrently": "^7.0.0",
+    "nodemon": "^2.0.15",
     "test": "0.6.0",
-    "tslint": "5.11.0",
-    "tslint-eslint-rules": "5.4.0",
-    "tslint-eslint-rules-recommended": "1.2.2",
-    "typescript": "4.2.3"
+    "typescript": "^4.6.2"
   }
 }
